lib files was moved to "library" target
